requirements.txt
snowflake-connector-python>=3.0.0
openai>=1.14.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
jinja2>=3.1.0
python-dotenv>=1.0.0
tenacity>=8.0.0
prometheus-client>=0.17.0
colorama>=0.4.0